+++ examples/plotting/Finding_Local_Peaks_in_Solar_Data.py
@@ -20,6 +20,7 @@ from skimage.feature import peak_local_max
 
 import sunpy.map
 from sunpy.data.sample import AIA_193_IMAGE
+from sunpy.map.maputils import all_pixel_indices_from_map
 
 ###############################################################################
 # We will first create a Map using some sample data and display it.
@@ -32,13 +33,9 @@ plt.colorbar()
 
 ###############################################################################
 # Before we find regions of local maxima, we need to create some variables to
-# store pixel coordinates for the 2D SDO/AIA data we are using.
+# store pixel values for the 2D SDO/AIA data we are using.
 # These variables are used for plotting in 3D later on.
-
-x = np.arange(aiamap.data.shape[0])
-y = np.arange(aiamap.data.shape[1])
-X, Y = np.meshgrid(x, y)
-
+X, Y = all_pixel_indices_from_map(aiamap)
 
 #######################################################################################
 # We will only consider peaks within the AIA data that have minimum intensity