+++ sunpy/net/dataretriever/sources/lyra.py
@@ -10,6 +10,30 @@ __all__ = ['LYRAClient']
 
 
 class LYRAClient(GenericClient):
+    """
+    Provides access to the LYRA/Proba2 data `archive <http://proba2.oma.be/lyra/data/bsd/>`__
+    hosted by the `PROBA2 Science Center <http://proba2.oma.be>`__.
+
+    Examples
+    --------
+
+    >>> from sunpy.net import Fido, attrs as a
+    >>> results = Fido.search(a.Time("2016/1/1", "2016/1/2"),
+    ...                       a.Instrument('LYRA'))  #doctest: +REMOTE_DATA
+    >>> results  #doctest: +REMOTE_DATA +ELLIPSIS
+    <sunpy.net.fido_factory.UnifiedResponse object at ...>
+    Results from 1 Provider:
+    <BLANKLINE>
+    2 Results from the LYRAClient:
+         Start Time           End Time      Source Instrument Wavelength
+           str19               str19         str6     str4       str3
+    ------------------- ------------------- ------ ---------- ----------
+    2016-01-01 00:00:00 2016-01-02 00:00:00 Proba2       lyra        nan
+    2016-01-01 00:00:00 2016-01-02 00:00:00 Proba2       lyra        nan
+    <BLANKLINE>
+    <BLANKLINE>
+
+    """
     def _get_url_for_timerange(self, timerange, **kwargs):
         """
         Returns list of URLS corresponding to value of input timerange.